requirements.txt
dotenv
streamlit
google-generativeai
python-dotenv
os
json